+++ src/main/java/pro/sky/animalsheltertelegrambot/service/impl/UserServiceImpl.java
@@ -5,6 +5,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import pro.sky.animalsheltertelegrambot.exception.UserNotFoundException;
 import pro.sky.animalsheltertelegrambot.model.User;
+import pro.sky.animalsheltertelegrambot.repository.AdoptionRepository;
 import pro.sky.animalsheltertelegrambot.repository.UserRepository;
 import pro.sky.animalsheltertelegrambot.service.UserService;
 
@@ -28,6 +29,7 @@ import static pro.sky.animalsheltertelegrambot.utils.MethodNameRetriever.getMeth
 @Slf4j
 public class UserServiceImpl implements UserService {
 
+    private final AdoptionRepository adoptionRepository;
     private final UserRepository userRepository;
 
     @Override
@@ -76,4 +78,30 @@ public class UserServiceImpl implements UserService {
             throw new UserNotFoundException(" User with id not found");
         }
     }
+
+    public void saveNewUser(Long userId, String userName) {
+        User user = new User(userId, userName, false);
+        userRepository.save(user);
+    }
+
+    /**
+     * Проверка на существование пользователя в БД.
+     *
+     * @param userId chat ID, который также является ID пользователя
+     * @return true - если пользователя еще нет в базе, false - если есть
+     */
+    public boolean checkIsUserIsNew(Long userId) {
+        return userRepository.findById(userId).isEmpty();
+    }
+
+    /**
+     * Производится проверка, является ли пользователь усыновителем или нет.
+     *
+     * @param userId chat ID, который также является ID пользователя
+     * @return true - если пользователь является усыновителем, false - если нет
+     */
+    public boolean checkIfUserIsAdopter(Long userId) {
+        return adoptionRepository.checkAdoptionsIsActive(userId).orElse(false);
+    }
+
 }
