+++ src/main/java/pro/sky/animalsheltertelegrambot/telegram_bot/service/CallbackService.java
@@ -1,0 +1,4 @@
+package pro.sky.animalsheltertelegrambot.telegram_bot.service;
+
+public interface CallbackService {
+}
