+++ src/main/java/pro/sky/animalsheltertelegrambot/telegram_bot/service/MessageSendingService.java
@@ -1,0 +1,6 @@
+package pro.sky.animalsheltertelegrambot.telegram_bot.service;
+
+public interface MessageSendingService {
+
+    public void sendMessage(Long chatId, String sendingMessage);
+}
