+++ src/main/java/pro/sky/animalsheltertelegrambot/telegram_bot/service/MessageService.java
@@ -1,0 +1,8 @@
+package pro.sky.animalsheltertelegrambot.telegram_bot.service;
+
+import com.pengrad.telegrambot.model.Message;
+
+public interface MessageService {
+
+    public void handleMessage(Message message);
+}
